+++ pycm/__init__.py
@@ -6,5 +6,6 @@ from .pycm_output import pycm_help, online_help
 from .pycm_distance import DistanceType
 from .pycm_obj import ConfusionMatrix
 from .pycm_compare import Compare
+from .pycm_multilabel_cm import MultiLabelCM
 from .pycm_curve import Curve, ROCCurve, PRCurve
 __version__ = PYCM_VERSION